requirements.txt
numpy
colorlog
ipython
tqdm
tensorflow-gpu==1.12.0
pillow
